+++ client/src/routes/iracing/tracks.$trackOrdinal.index.tsx
@@ -1,6 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-import { TrackDetailRoute } from "../../components/track/TrackDetailRoute";
-
-export const Route = createFileRoute("/iracing/tracks/$trackOrdinal/")({
-  component: () => <TrackDetailRoute tab="info" />,
-});